Refactor to move loadFake into its own file
